main_app/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.7
isodate==0.6.1
loguru==0.7.2
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
platformdirs==4.2.0
py-machineid==0.5.1
pydantic==2.6.4
pydantic_core==2.16.3
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.3
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
setuptools==69.2.0
shiboken6==6.7.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
win32-setctime==1.1.0
winregistry==1.1.1
zeep==4.2.1
